Worked on Feedback changes, including removal of package-lock.json
