--- src/components/header/Header.js
@@ -1,28 +1,0 @@
-import React from "react";
-import "./header.css";
-
-const Header = () => {
-  return (
-    <header className="header">
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        clip-rule="evenodd"
-        fill-rule="evenodd"
-        stroke-linejoin="round"
-        stroke-miterlimit="2"
-        viewBox="0 0 24 24"
-        width="40"
-        height="40"
-        fill="#fff"
-      >
-        <path
-          d="m20 20h-15.25c-.414 0-.75.336-.75.75s.336.75.75.75h15.75c.53 0 1-.47 1-1v-15.75c0-.414-.336-.75-.75-.75s-.75.336-.75.75zm-1-17c0-.478-.379-1-1-1h-15c-.62 0-1 .519-1 1v15c0 .621.52 1 1 1h15c.478 0 1-.379 1-1zm-12.751 8.306c-.165-.147-.249-.351-.249-.556 0-.411.333-.746.748-.746.178 0 .355.062.499.19l2.371 2.011 4.453-4.962c.149-.161.35-.243.554-.243.417 0 .748.336.748.746 0 .179-.065.359-.196.503l-4.953 5.508c-.148.161-.35.243-.553.243-.177 0-.356-.063-.498-.19z"
-          fill-rule="nonzero"
-        ></path>
-      </svg>
-      <h1 className="title">TODO</h1>
-    </header>
-  );
-};
-
-export default Header;
