--- src/components/todoForm/TodoForm.js
@@ -1,51 +1,0 @@
-import React from "react";
-import { useState } from "react";
-import "./todoform.css";
-
-const TodoForm = (props) => {
-  const [input, setInput] = useState("");
-
-  const inputChangeHandler = (event) => {
-    setInput(event.target.value);
-  };
-
-  const submitHandler = (event) => {
-    event.preventDefault();
-    props.submitHandler({ text: input });
-    setInput("");
-  };
-
-  return (
-    <form className="form" onSubmit={submitHandler}>
-      <label for="todo">
-        <input
-          id="todo"
-          type="text"
-          onChange={inputChangeHandler}
-          value={input}
-          placeholder="Write your next task"
-        />
-      </label>
-      <button type="submit">
-        <span class="visually-hidden">Submit</span>
-        <svg
-          clip-rule="evenodd"
-          fill-rule="evenodd"
-          stroke-linejoin="round"
-          stroke-miterlimit="2"
-          viewBox="0 0 24 24"
-          xmlns="http://www.w3.org/2000/svg"
-          width="32"
-          height="32"
-        >
-          <path
-            d="m11 11h-7.25c-.414 0-.75.336-.75.75s.336.75.75.75h7.25v7.25c0 .414.336.75.75.75s.75-.336.75-.75v-7.25h7.25c.414 0 .75-.336.75-.75s-.336-.75-.75-.75h-7.25v-7.25c0-.414-.336-.75-.75-.75s-.75.336-.75.75z"
-            fill-rule="nonzero"
-          ></path>
-        </svg>
-      </button>
-    </form>
-  );
-};
-
-export default TodoForm;
