+++ src/shared/Button.js
@@ -1,0 +1,19 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const Button = ({ type = "button", onClick, children, className }) => {
+  return (
+    <button type={type} onClick={onClick} className={className}>
+      {children}
+    </button>
+  );
+};
+
+Button.propTypes = {
+  type: PropTypes.string,
+  onClick: PropTypes.func,
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string,
+};
+
+export default Button;
